+++ src/migratus/database.clj
@@ -290,7 +290,8 @@
   proto/Store
   (config [this] config)
   (init [this]
-    (let [conn (connect* (assoc (:db config) :transaction? (:init-in-transaction? config)))]
+    (let [conn (connect* (assoc (:db config)
+                                :transaction? (:init-in-transaction? config)))]
       (try
         (init-db! conn
                   (utils/get-migration-dir config)
@@ -305,24 +306,25 @@
   (completed [this]
     (completed* @connection (migration-table-name config)))
   (migrate-up [this migration]
-              (log/info "Connection is " @connection
-                        "Config is" (update config :db utils/censor-password))
-              (if (proto/tx? migration :up)
-                (jdbc/with-transaction [t-con (connection-or-spec @connection)]
-                  (migrate-up* t-con config migration))
-                (migrate-up* (:db config) config migration)))
+    (log/info "Connection is " @connection
+              "Config is" (update config :db utils/censor-password))
+    (if (proto/tx? migration :up)
+      (jdbc/with-transaction [t-con (connection-or-spec @connection)]
+        (migrate-up* t-con config migration))
+      (migrate-up* (:db config) config migration)))
   (migrate-down [this migration]
-                (log/info "Connection is " @connection
-                          "Config is" (update config :db utils/censor-password))
-                (if (proto/tx? migration :down)
-                  (jdbc/with-transaction [t-con (connection-or-spec @connection)]
-                    (migrate-down* t-con config migration))
-                  (migrate-down* (:db config) config migration)))
+    (log/info "Connection is " @connection
+              "Config is" (update config :db utils/censor-password))
+    (if (proto/tx? migration :down)
+      (jdbc/with-transaction [t-con (connection-or-spec @connection)]
+        (migrate-down* t-con config migration))
+      (migrate-down* (:db config) config migration)))
   (connect [this]
     (reset! connection (connect* (:db config)))
     (init-schema! @connection
                   (migration-table-name config)
-                  (sql-mig/wrap-modify-sql-fn (:modify-sql-fn config))))
+                  (sql-mig/wrap-modify-sql-fn (:modify-sql-fn config)))
+    this)
   (disconnect [this]
     (disconnect* @connection config)
     (reset! connection nil)))